escuela-fastapi/requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
psycopg2-binary
boto3
python-multipart
requests
python-dotenv